v2.3.0
详情见gitee

+++ DrissionPage/drission.py
@@ -15,15 +15,14 @@ from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.chrome.webdriver import WebDriver
 from tldextract import extract
 
-from .config import (_dict_to_chrome_options, _session_options_to_dict,
-                     SessionOptions, DriverOptions, _chrome_options_to_dict, OptionsManager, _cookies_to_tuple)
+from .config import _session_options_to_dict, SessionOptions, DriverOptions, _cookies_to_tuple
 
 
 class Drission(object):
     """Drission类用于管理WebDriver对象和Session对象，是驱动器的角色"""
 
     def __init__(self,
-                 driver_or_options: Union[WebDriver, dict, Options, DriverOptions, bool] = None,
+                 driver_or_options: Union[WebDriver, Options, DriverOptions, bool] = None,
                  session_or_options: Union[Session, dict, SessionOptions, bool] = None,
                  ini_path: str = None,
                  proxy: dict = None):
@@ -35,37 +34,45 @@ class Drission(object):
         """
         self._session = None
         self._driver = None
+        self._session_options = None
+        self._driver_options = None
         self._debugger = None
         self._proxy = proxy
 
-        om = OptionsManager(ini_path) if session_or_options is None or driver_or_options is None else None
-
         # ------------------处理session options----------------------
         if session_or_options is None:
-            self._session_options = om.session_options
+            self._session_options = SessionOptions(ini_path=ini_path).as_dict()
+
+        elif session_or_options is False:
+            self._driver_options = SessionOptions(read_file=False).as_dict()
+
+        elif isinstance(session_or_options, Session):
+            self._session = session_or_options
+
+        elif isinstance(session_or_options, SessionOptions):
+            self._session_options = session_or_options.as_dict()
+
+        elif isinstance(session_or_options, dict):
+            self._session_options = session_or_options
 
         else:
-            # 若接收到Session对象，直接记录
-            if isinstance(session_or_options, Session):
-                self._session = session_or_options
-
-            # 否则记录其配置信息
-            else:
-                self._session_options = _session_options_to_dict(session_or_options)
+            raise TypeError('session_or_options参数只能接收Session, dict, SessionOptions或False。')
 
         # ------------------处理driver options----------------------
         if driver_or_options is None:
-            self._driver_options = om.chrome_options
-            self._driver_options['driver_path'] = om.get_value('paths', 'chromedriver_path')
+            self._driver_options = DriverOptions(ini_path=ini_path)
+
+        elif driver_or_options is False:
+            self._driver_options = DriverOptions(read_file=False)
+
+        elif isinstance(driver_or_options, WebDriver):
+            self._driver = driver_or_options
+
+        elif isinstance(driver_or_options, (Options, DriverOptions)):
+            self._driver_options = driver_or_options
 
         else:
-            # 若接收到WebDriver对象，直接记录
-            if isinstance(driver_or_options, WebDriver):
-                self._driver = driver_or_options
-
-            # 否则记录其配置信息
-            else:
-                self._driver_options = _chrome_options_to_dict(driver_or_options)
+            raise TypeError('driver_or_options参数只能接收WebDriver, Options, DriverOptions或False。')
 
     def __del__(self):
         """关闭对象时关闭浏览器和Session"""
@@ -91,28 +98,23 @@ class Drission(object):
         如设置了本地调试浏览器，可自动接入或打开浏览器进程。
         """
         if self._driver is None:
-            if not isinstance(self._driver_options, dict):
-                raise TypeError('无效的Driver配置。')
+            if not self.driver_options.debugger_address and self._proxy:
+                self.driver_options.add_argument(f'--proxy-server={self._proxy["http"]}')
 
-            options = _dict_to_chrome_options(self._driver_options)
-
-            if not self._driver_options.get('debugger_address', None) and self._proxy:
-                options.add_argument(f'--proxy-server={self._proxy["http"]}')
-
-            driver_path = self._driver_options.get('driver_path', None) or 'chromedriver'
-            chrome_path = self._driver_options.get('binary_location', None) or 'chrome.exe'
+            driver_path = self.driver_options.driver_path or 'chromedriver'
+            chrome_path = self.driver_options.binary_location or 'chrome.exe'
 
             # -----------若指定debug端口且该端口未在使用中，则先启动浏览器进程-----------
-            if options.debugger_address and _check_port(options.debugger_address) is False:
+            if self.driver_options.debugger_address and _check_port(self.driver_options.debugger_address) is False:
                 from subprocess import Popen
-                port = options.debugger_address.split(':')[-1]
+                port = self.driver_options.debugger_address.split(':')[-1]
 
                 # 启动浏览器进程，同时返回该进程使用的 chrome.exe 路径
                 chrome_path, self._debugger = _create_chrome(chrome_path, port,
-                                                             self._driver_options['arguments'], self._proxy)
+                                                             self.driver_options.arguments, self._proxy)
 
             # -----------创建WebDriver对象-----------
-            self._driver = _create_driver(chrome_path, driver_path, options)
+            self._driver = _create_driver(chrome_path, driver_path, self.driver_options)
 
             # 反反爬设置
             try:
@@ -127,7 +129,7 @@ class Drission(object):
         return self._driver
 
     @property
-    def driver_options(self) -> dict:
+    def driver_options(self) -> Union[DriverOptions, Options]:
         """返回driver配置信息"""
         return self._driver_options
 
@@ -185,18 +187,21 @@ class Drission(object):
 
         pid = self.get_browser_progress_id()
         from os import popen
-        if pid and popen(f'tasklist | findstr {pid}').read().lower().startswith('chrome.exe'):
-            print('kill')
+        from platform import system
+
+        if pid and system().lower() == 'windows' \
+                and popen(f'tasklist | findstr {pid}').read().lower().startswith('chrome.exe'):
             popen(f'taskkill /pid {pid} /F')
+
         else:
-            self.close_driver()
+            self._driver.quit()
 
     def get_browser_progress_id(self) -> Union[str, None]:
         """获取浏览器进程id"""
         if self.debugger_progress:
             return self.debugger_progress.pid
 
-        address = self.driver_options.get('debugger_address', '').split(':')
+        address = str(self.driver_options.debugger_address).split(':')
         if len(address) == 2:
             ip, port = address
             if ip not in ('127.0.0.1', 'localhost') or not port.isdigit():
@@ -223,6 +228,10 @@ class Drission(object):
         self._show_or_hide_browser(False)
 
     def _show_or_hide_browser(self, hide: bool = True) -> None:
+        from platform import system
+        if system().lower() != 'windows':
+            raise OSError('该方法只能在Windows系统使用。')
+
         try:
             from win32gui import ShowWindow
             from win32con import SW_HIDE, SW_SHOW
@@ -360,10 +369,14 @@ class Drission(object):
         selenium_user_agent = driver.execute_script("return navigator.userAgent;")
         session.headers.update({"User-Agent": selenium_user_agent})
 
-    def close_driver(self) -> None:
+    def close_driver(self, kill: bool = False) -> None:
         """关闭driver和浏览器"""
         if self._driver:
-            self._driver.quit()
+            if kill:
+                self.kill_browser()
+            else:
+                self._driver.quit()
+
             self._driver = None
 
     def close_session(self) -> None:
@@ -434,7 +447,7 @@ def _create_chrome(chrome_path: str, port: str, args: list, proxy: dict) -> tupl
         debugger = Popen(f'{chrome_path} --remote-debugging-port={port} {args}', shell=False)
 
         if chrome_path == 'chrome.exe':
-            from common import get_exe_path_from_port
+            from .common import get_exe_path_from_port
             chrome_path = get_exe_path_from_port(port)
 
     # 传入的路径找不到，主动在ini文件、注册表、系统变量中找
@@ -458,29 +471,27 @@ def _create_driver(chrome_path: str, driver_path: str, options: Options) -> WebD
     :return: WebDriver 对象
     """
     try:
+        debugger_address = options.debugger_address
+        if options.debugger_address:
+            options = Options()
+            options.debugger_address = debugger_address
+
         return webdriver.Chrome(driver_path, options=options)
 
     # 若版本不对，获取对应 chromedriver 再试
     except (WebDriverException, SessionNotCreatedException):
+        print('打开失败，尝试获取driver。\n')
         from .easy_set import get_match_driver
-        chrome_path = None if chrome_path == 'chrome.exe' else chrome_path
-        driver_path = get_match_driver(chrome_path=chrome_path, check_version=False, show_msg=False)
+        from DrissionPage.easy_set import _get_chrome_path
 
-        if driver_path:
-            try:
-                return webdriver.Chrome(driver_path, options=options)
-            except Exception:
-                pass
+        if chrome_path == 'chrome.exe':
+            chrome_path = _get_chrome_path(show_msg=False, from_ini=False)
 
-        # 当找不到 driver 且 chrome_path 为 None 时，说明安装的版本过高，改在系统路径中查找
-        elif chrome_path is None and driver_path is None:
-            from DrissionPage.easy_set import _get_chrome_path
-            chrome_path = _get_chrome_path(show_msg=False, from_ini=False, from_regedit=False)
-            driver_path = get_match_driver(chrome_path=chrome_path, check_version=False, show_msg=False)
-
+        if chrome_path:
+            driver_path = get_match_driver(chrome_path=chrome_path, check_version=False, show_msg=True)
             if driver_path:
-                options.binary_location = chrome_path
                 try:
+                    options.binary_location = chrome_path
                     return webdriver.Chrome(driver_path, options=options)
                 except Exception:
                     pass
